HealthAndStat v1.0
Items now changes the values in Stat and Health Managers and some of those stats are already affecting different outcomes. Others like damage multipliers are ready to be attached to combat system once one is made

+++ Assets/Scripts/Enviroment/Spikes.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 namespace Environment {
     public class Spikes : MonoBehaviour {
-        [SerializeField] private int damage;
+        [SerializeField] private float damage;
 
         private void OnTriggerStay2D(Collider2D other) {
             if (other.CompareTag("Player") && other.TryGetComponent(out HealthManager healthManager)) {
